Set up Add Another Lesson functionality
